requirements.txt
anyio==4.2.0
asttokens==2.4.1
blinker==1.8.2
blurhash==1.1.4
bsky-bridge==1.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
commonmark==0.9.1
decorator==5.1.1
docutils==0.21.2
exceptiongroup==1.2.0
executing==2.1.0
feedgenerator==2.1.0
feedparser==6.0.11
idna==3.6
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
Mastodon.py==1.8.1
matplotlib-inline==0.1.7
mdurl==0.1.2
nh3==0.2.18
oauthlib==3.2.2
ordered-set==4.1.0
parso==0.8.4
pelican==4.10.1
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.0
ruff==0.8.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tomli==2.1.0
traitlets==5.14.3
tweepy==4.14.0
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.3
watchfiles==0.21.0
wcwidth==0.2.13
